+++ wow_world_base/src/inner/vanilla/battleground_bracket.rs
@@ -4,7 +4,7 @@
 ///     TENS = 0;
 ///     TWENTIES = 1;
 ///     THIRTIES = 2;
-///     FOURTIES = 3;
+///     FORTIES = 3;
 ///     FIFTIES = 4;
 ///     SIXTY = 5;
 /// }
@@ -19,7 +19,7 @@ pub enum BattlegroundBracket {
     /// 30-39
     Thirties,
     /// 40-49
-    Fourties,
+    Forties,
     /// 50-59
     Fifties,
     /// 60
@@ -32,7 +32,7 @@ impl BattlegroundBracket {
             Self::Tens => 0x0,
             Self::Twenties => 0x1,
             Self::Thirties => 0x2,
-            Self::Fourties => 0x3,
+            Self::Forties => 0x3,
             Self::Fifties => 0x4,
             Self::Sixty => 0x5,
         }
@@ -43,7 +43,7 @@ impl BattlegroundBracket {
             Self::Tens,
             Self::Twenties,
             Self::Thirties,
-            Self::Fourties,
+            Self::Forties,
             Self::Fifties,
             Self::Sixty,
         ]
@@ -54,7 +54,7 @@ impl BattlegroundBracket {
             0 => Ok(Self::Tens),
             1 => Ok(Self::Twenties),
             2 => Ok(Self::Thirties),
-            3 => Ok(Self::Fourties),
+            3 => Ok(Self::Forties),
             4 => Ok(Self::Fifties),
             5 => Ok(Self::Sixty),
             v => Err(crate::errors::EnumError::new(NAME, v as i128),)
@@ -69,7 +69,7 @@ impl BattlegroundBracket {
             Self::Tens => "TENS",
             Self::Twenties => "TWENTIES",
             Self::Thirties => "THIRTIES",
-            Self::Fourties => "FOURTIES",
+            Self::Forties => "FORTIES",
             Self::Fifties => "FIFTIES",
             Self::Sixty => "SIXTY",
         }
@@ -91,7 +91,7 @@ impl std::fmt::Display for BattlegroundBracket {
             Self::Tens => f.write_str("Tens"),
             Self::Twenties => f.write_str("Twenties"),
             Self::Thirties => f.write_str("Thirties"),
-            Self::Fourties => f.write_str("Fourties"),
+            Self::Forties => f.write_str("Forties"),
             Self::Fifties => f.write_str("Fifties"),
             Self::Sixty => f.write_str("Sixty"),
         }